Convert2DTo3D_new

+++ SI/MyApplication/app/src/main/java/com/example/si/IMG_PROCESSING/CornerDetection/RGB32f.java
@@ -12,9 +12,9 @@ public class RGB32f implements Cloneable{
     }
 
     public RGB32f(XYZ a){
-        r = a.x;
-        g = a.y;
-        b = a.z;
+        r = (float) a.x;
+        g = (float) a.y;
+        b = (float) a.z;
     }
 
     @Override
